requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
Pydantic==2.5.2
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
